Snac updates/reorganization, loader improvements, bug fixes
